move message-related things to message.rs
