+++ src/wrapper.rs
@@ -1,73 +1,10 @@
-use std::cell::UnsafeCell;
-use ringbuf::{RingBuffer, Producer, Consumer};
+use ringbuf::RingBuffer;
 
-use crate::{AudioBus, AudioBusMut, Event, MidiReceiver, Model, MusicalTime, Param, Parameters, Plugin, PluginUI, ProcessContext, SmoothModel, event};
-
-pub trait UIHostCallback: Send + Sync {
-    fn send_parameter_update(&self, param_idx: usize, normalized: f32);
-
-    // Called when the UI Model received the `ShouldClose` message.
-    fn close_msg_received(&self);
-}
-
-pub enum PlugToUIMsg<Model: 'static> {
-    ParamChanged {
-        // Sending parameter by index because I cannot for the life of me figure out how to make
-        // the rust compiler happy with all the generics.
-        param_idx: usize,
-        normalized: f32,
-    },
-    ProgramChanged(Box<Model>),
-    ShouldClose,
-}
-
-pub enum UIToPlugMsg<SmoothModel: 'static> {
-    ParamChanged {
-        // Sending parameter by index because I cannot for the life of me figure out how to make
-        // the rust compiler happy with all the generics.
-        param_idx: usize,
-        normalized: f32,
-    },
-    ValueChanged {
-        // Holy crap this actually works.
-        cb: &'static fn(&mut SmoothModel, f32),
-        value: f32,
-    },
-    Closed
-}
-
-pub(crate) struct UIMsgHandles<P: Plugin> {
-    pub plug_to_ui_tx: Producer<PlugToUIMsg<P::Model>>,
-    pub ui_to_plug_rx: Consumer<UIToPlugMsg<<P::Model as Model<P>>::Smooth>>,
-}
-
-pub struct PlugMsgHandles<Model: 'static, SmoothModel: 'static> {
-    pub ui_host_cb: Box<dyn UIHostCallback>,
-    pub notify_dsp: bool,
-
-    plug_to_ui_rx: UnsafeCell<Consumer<PlugToUIMsg<Model>>>,
-    ui_to_plug_tx: UnsafeCell<Producer<UIToPlugMsg<SmoothModel>>>,
-}
-
-impl<Model: 'static, SmoothModel: 'static> PlugMsgHandles<Model, SmoothModel> {
-    pub fn pop_msg(&self) -> Option<PlugToUIMsg<Model>> {
-        // Safe because this is only place this is borrowed, and this is just a message queue.
-        unsafe { (&mut *self.plug_to_ui_rx.get()).pop() }
-    }
-
-    pub fn push_msg(&self, msg: UIToPlugMsg<SmoothModel>) -> Result<(), UIToPlugMsg<SmoothModel>> {
-        // Safe because this is only place this is borrowed, and this is just a message queue.
-        unsafe { (&mut *self.ui_to_plug_tx.get()).push(msg) }
-    }
-}
-
-impl<Model: 'static, SmoothModel: 'static> Drop for PlugMsgHandles<Model, SmoothModel> {
-    fn drop(&mut self) {
-        if let Err(_) = self.push_msg(UIToPlugMsg::Closed) {
-            eprintln!("UI to Plug message buffer is full!");
-        }
-    }
-}
+use crate::{
+    AudioBus, AudioBusMut, Event, MidiReceiver, Model, MusicalTime, Param, Parameters,
+    Plugin, PluginUI, PlugToUIMsg, PlugMsgHandles, ProcessContext, SmoothModel,
+    UIMsgHandles, UIToPlugMsg, UIHostCallback, event
+};
 
 pub(crate) struct WrappedPlugin<P: Plugin> {
     pub(crate) plug: P,
@@ -221,12 +158,12 @@ impl<P: Plugin> WrappedPlugin<P> {
 
         let model = self.smoothed_model.as_model();
 
-        let plug_msg_handles = PlugMsgHandles {
-            ui_host_cb: ui_host_callback,
-            plug_to_ui_rx: UnsafeCell::new(plug_to_ui_rx),
-            ui_to_plug_tx: UnsafeCell::new(ui_to_plug_tx),
+        let plug_msg_handles = PlugMsgHandles::new(
+            ui_host_callback,
+            plug_to_ui_rx,
+            ui_to_plug_tx,
             notify_dsp,
-        };
+        );
 
         <<P::Model as Model<P>>::UI as UIModel<P, P::Model>>::from_model(
             model,
